refactored folder structure according to what we learned
